+++ integration-test/smoke/helix-smoke-test-cookies.json
@@ -1,0 +1,1 @@
+{"url":"https://app.tryhelix.ai","cookies":[{"name":"rl_user_id","value":"RS_ENC_v3_InBoaWwrc21va2V0ZXN0QGhlbGl4Lm1sIg%3D%3D","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1766227331},{"name":"rl_trait","value":"RS_ENC_v3_eyJpZCI6ImQzY2NiNzdiLTYxYTAtNGIzYS05MGQyLThmOTIwYzVmYTU2MSIsImVtYWlsIjoicGhpbCtzbW9rZXRlc3RAaGVsaXgubWwiLCJuYW1lIjoiU21va2UgVGVzdCJ9","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1766227331},{"name":"rl_session","value":"RS_ENC_v3_eyJpZCI6MTczNDY5MTMzMDg0NiwiZXhwaXJlc0F0IjoxNzM0NjkzMTMxOTUxLCJ0aW1lb3V0IjoxODAwMDAwLCJhdXRvVHJhY2siOnRydWUsInNlc3Npb25TdGFydCI6dHJ1ZX0%3D","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1766227331},{"name":"_clck","value":"czxqsb%7C2%7Cfrv%7C1%7C1815","domain":".tryhelix.ai","path":"/","expires":1766227331},{"name":"AUTH_SESSION_ID_LEGACY","value":"ac21117a-e5a7-4208-95bb-ab91a3bf5817.c32817ec5e68-10112","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","httpOnly":true,"expires":-1},{"name":"_ga_TMZRHCF42M","value":"GS1.1.1734691329.1.1.1734691331.0.0.0","domain":".tryhelix.ai","path":"/","expires":1769251331.509285},{"name":"KEYCLOAK_SESSION_LEGACY","value":"helix/d3ccb77b-61a0-4b3a-90d2-8f920c5fa561/ac21117a-e5a7-4208-95bb-ab91a3bf5817","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","expires":1742467331.203051},{"name":"KEYCLOAK_SESSION","value":"helix/d3ccb77b-61a0-4b3a-90d2-8f920c5fa561/ac21117a-e5a7-4208-95bb-ab91a3bf5817","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","secure":true,"sameSite":"None","expires":1742467331.203007},{"name":"_clsk","value":"wtuvpu%7C1734691332273%7C2%7C1%7Cj.clarity.ms%2Fcollect","domain":".tryhelix.ai","path":"/","expires":1734777732},{"name":"KEYCLOAK_IDENTITY_LEGACY","value":"eyJhbGciOiJIUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICI4NjQ4YzVmZC1mOWNiLTQ1ZDUtYWFhNy03ZWY0ODk5MDBhZGEifQ.eyJleHAiOjE3NDI0NjczMzEsImlhdCI6MTczNDY5MTMzMSwianRpIjoiMWYwMjAyODEtYTQ1Yy00OTJhLTk4YWEtYTk1MDRmMDliMDM0IiwiaXNzIjoiaHR0cHM6Ly9hcHAudHJ5aGVsaXguYWkvYXV0aC9yZWFsbXMvaGVsaXgiLCJzdWIiOiJkM2NjYjc3Yi02MWEwLTRiM2EtOTBkMi04ZjkyMGM1ZmE1NjEiLCJ0eXAiOiJTZXJpYWxpemVkLUlEIiwic2Vzc2lvbl9zdGF0ZSI6ImFjMjExMTdhLWU1YTctNDIwOC05NWJiLWFiOTFhM2JmNTgxNyIsInNpZCI6ImFjMjExMTdhLWU1YTctNDIwOC05NWJiLWFiOTFhM2JmNTgxNyIsInN0YXRlX2NoZWNrZXIiOiJCWXNjS3E4VUV5X2NNb2RKelBydWwwYnF0R3F3U0xFTHF0LVItN3QwMXFNIn0.hye4juU6skvoRl7Y3AJc4tK2MSKb2B9rIDi4jyprV9I","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","httpOnly":true,"expires":-1},{"name":"KEYCLOAK_IDENTITY","value":"eyJhbGciOiJIUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICI4NjQ4YzVmZC1mOWNiLTQ1ZDUtYWFhNy03ZWY0ODk5MDBhZGEifQ.eyJleHAiOjE3NDI0NjczMzEsImlhdCI6MTczNDY5MTMzMSwianRpIjoiMWYwMjAyODEtYTQ1Yy00OTJhLTk4YWEtYTk1MDRmMDliMDM0IiwiaXNzIjoiaHR0cHM6Ly9hcHAudHJ5aGVsaXguYWkvYXV0aC9yZWFsbXMvaGVsaXgiLCJzdWIiOiJkM2NjYjc3Yi02MWEwLTRiM2EtOTBkMi04ZjkyMGM1ZmE1NjEiLCJ0eXAiOiJTZXJpYWxpemVkLUlEIiwic2Vzc2lvbl9zdGF0ZSI6ImFjMjExMTdhLWU1YTctNDIwOC05NWJiLWFiOTFhM2JmNTgxNyIsInNpZCI6ImFjMjExMTdhLWU1YTctNDIwOC05NWJiLWFiOTFhM2JmNTgxNyIsInN0YXRlX2NoZWNrZXIiOiJCWXNjS3E4VUV5X2NNb2RKelBydWwwYnF0R3F3U0xFTHF0LVItN3QwMXFNIn0.hye4juU6skvoRl7Y3AJc4tK2MSKb2B9rIDi4jyprV9I","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","secure":true,"httpOnly":true,"sameSite":"None","expires":-1},{"name":"rl_page_init_referrer","value":"RS_ENC_v3_IiRkaXJlY3Qi","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1766227331},{"name":"rl_anonymous_id","value":"RS_ENC_v3_ImMzN2RmYzlmLTI3MjQtNDU3YS04NzY2LWViZGFkODNjMjdkNiI%3D","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1766227331},{"name":"crisp-client%2Fsession%2Fd69e5955-fa6a-4fe6-b1bb-c87cf7515d09","value":"session_91ddbdd5-48d8-48c3-acdc-1edf19a9cb9d","domain":".tryhelix.ai","path":"/","sameSite":"Lax","expires":1750459331},{"name":"_ga","value":"GA1.1.817628766.1734691330","domain":".tryhelix.ai","path":"/","expires":1769251331.50533},{"name":"AUTH_SESSION_ID","value":"ac21117a-e5a7-4208-95bb-ab91a3bf5817.c32817ec5e68-10112","domain":"app.tryhelix.ai","path":"/auth/realms/helix/","secure":true,"httpOnly":true,"sameSite":"None","expires":-1}],"time":"2024-12-20T10:42:13.350682Z"}
